+++ prisma/migrations/20221022212118_add_url_index/migration.sql
@@ -1,0 +1,2 @@
+-- CreateIndex
+CREATE INDEX "Link_url_userId_idx" ON "Link"("url", "userId");